refactor(cli): extract sandbox doctor action
